tutorial/requirements.txt
Django==1.10.2
djangorestframework==3.4.7
httpie==0.9.6
Pygments==2.1.3
requests==2.11.1
